requirements.txt
panda3d
DirectGuiExtension
